requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-dotenv>=1.0
requests>=2.31

# Data processing
pandas>=2.1
numpy>=1.24
pyarrow>=14.0

# ML and MLOps
scikit-learn>=1.3
mlflow>=2.9
joblib>=1.3

# AWS and S3
boto3>=1.28
s3fs>=2023.9

# Workflow and scheduling
prefect>=2.14
schedule>=1.2

# Monitoring and logging
wandb>=0.16
psutil>=5.9
pytz>=2023.3

# Database
redis>=5.0

# Additional utilities for production
sentry-sdk[fastapi]>=1.38
prometheus-client>=0.19

# Workflow orchestration
apache-airflow>=2.8.0
apache-airflow-providers-postgres>=5.7.0
apache-airflow-providers-amazon>=8.11.0